services/ml-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-multipart==0.0.9
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
joblib==1.4.0
matplotlib==3.8.4
